Factor out expression type checking into a separate file. No functionality change
